chore: move file rediret-to
