+++ app/api/auth/agent-connect/logout-callback/route.ts
@@ -1,7 +1,7 @@
 import logErrorInSentry from '#utils/sentry';
+import { redirectTo } from '#utils/server-side-helper/app/redirect-to';
 import { cleanAgentSession, cleanPathFrom, getPathFrom } from '#utils/session';
 import withSession from '#utils/session/with-session';
-import { redirectTo } from '../../utils';
 import { AgentConnectLogoutFailedException } from '../agent-connect-types';
 
 export const GET = withSession(async function logoutCallbackRoute(req) {
