+++ app/api/auth/agent-connect/logout/route.ts
@@ -1,9 +1,9 @@
 import { agentConnectLogoutUrl } from '#clients/authentication/agent-connect/strategy';
 import logErrorInSentry from '#utils/sentry';
+import { redirectTo } from '#utils/server-side-helper/app/redirect-to';
 import { setPathFrom } from '#utils/session';
 import withSession from '#utils/session/with-session';
 import { NextResponse } from 'next/server';
-import { redirectTo } from '../../utils';
 import { AgentConnectLogoutFailedException } from '../agent-connect-types';
 
 export const GET = withSession(async function logoutRoute(req) {
