+++ app/api/auth/france-connect/login/route.ts
@@ -1,8 +1,8 @@
 import { FranceConnectAuthorizeUrl } from '#clients/authentication/france-connect/strategy';
 import { logFatalErrorInSentry } from '#utils/sentry';
+import { redirectTo } from '#utils/server-side-helper/app/redirect-to';
 import withSession from '#utils/session/with-session';
 import { NextResponse } from 'next/server';
-import { redirectTo } from '../../utils';
 import { FranceConnectFailedException } from '../france-connect-types';
 
 export const GET = withSession(async function loginRoute(req) {
